improved rendering performance by using vertex buffer instead of buffer array and fixed breaking sync

+++ Shared/blocks/blocks.hpp
@@ -64,8 +64,8 @@ public:
     
     void breakBlock(int x, int y);
     
-    unsigned short getWidth() const;
-    unsigned short getHeight() const;
+    int getWidth() const;
+    int getHeight() const;
     
     void serialize(std::vector<char>& serial);
     char* loadFromSerial(char* iter);
